Crate fxbox_thinkerbell

This create defines mechanisms for executing simple scripts on the server.

By design, these scripts have limited capabilities. Each script takes the form of a set of rules: "when any of the input services of foo matches some condition, send some value to all of the onput services of bar".

See module ast for more details on the grammar of scripts.

Modules

ast
Definition of the AST. Definition of Thinkerbell scripts.

compile
Compiling an AST into something runnable. A script compiler

parse
Parsing JSON into an AST. Parsing scripts.

run
Actually executing code. Launching and running the script

util
Miscellaneous internal utilities.