Module fxbox_thinkerbell::ast

Definition of the AST. Definition of Thinkerbell scripts.

Typical applications will not interact with script objects. Rather, they will use module parse to parse a script and module run to execute it.

Structs

Match
An individual match.

Rule
A single rule, i.e. "when some condition becomes true, do something".

Script
A thinkerbell script.

Statement
Stuff to actually do. In practice, this means placing calls to devices.

UncheckedCtx
A Context used to represent a script that hasn't been compiled yet.

Traits

Context
A manner of representing internal nodes.