Module fxbox_thinkerbell::compile

Compiling an AST into something runnable. A script compiler

This compiler take untrusted code (Script<UncheckedCtx>) and performs the following transformations and checks:

Ensure that the Script has at least one Rule.

Ensure that each Rule has at least one Match.

Ensure that each Rule has at least one Statement.

Ensure that each Match has at least one source.

Ensure that each Statement has at least one destination.

Ensure that in each Match, the type of range matches the kind.

Ensure that in each Statement, the type of value matches the kind.

Transform each Match to make sure that the kind of the source matches the kind, even if devices change.

Transform each Statement to make sure that the kind of the destination matches the kind, even if devices change.

Structs

CompiledCtx

Compiler

Enums

Error

SourceError

TypeError

Traits

ExecutableDevEnv
The environment in which the code is meant to be executed. This can typically be instantiated either with actual bindings to devices, or with a unit-testing framework. // FIXME: Move this to run.rs